+++ src/main/java/app/mobile/BK_Kindergarten/attend/dto/CreateAttendRequest.java
@@ -5,7 +5,7 @@ import java.time.LocalDate;
 
 @Data
 public class CreateAttendRequest {
-    private Integer childId;  // ID học sinh được điểm danh
+    private Long childId;  // ID học sinh được điểm danh
     private Integer classId;  // ID lớp học
     private LocalDate date;   // Ngày điểm danh (nếu không gửi sẽ lấy ngày hiện tại)
     private String status;    // Trạng thái điểm danh (ví dụ: "Có mặt", "Vắng mặt")
